% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adjusted_ahsw}
\alias{adjusted_ahsw}
\title{Adjusted Average Hazard with Survival Weight (AHSW)}
\usage{
adjusted_ahsw(adjsurv, to, from=0, conf_int=FALSE, conf_level=0.95)
}
\arguments{
\item{adjsurv}{A list object resulting from an adjusted survival analysis
containing survival probabilities (`adjsurv$adj`) and optionally bootstrap
data (`adjsurv$boot_adj`).}

\item{to}{Numeric. The upper time limit (tau) for calculating RMST and AHSW.}

\item{from}{Numeric. Lower time limit for integration (default 0).}

\item{conf_int}{Logical. If TRUE, calculates bootstrap-based confidence intervals.}

\item{conf_level}{Numeric. Confidence level for intervals (default 0.95).}
}
\value{
A list containing:
\item{est}{A data frame with point estimates of RMST, event rate, and AHSW for two groups, their difference, and log difference.}
\item{rmst}{A data frame with bootstrap-based estimates, standard errors, confidence intervals, and p-values for RMST.}
\item{evrt}{Similar results as `rmst` but for event rate.}
\item{ahsw}{Similar results as `rmst` but for AHSW.}
}
\description{
Calculate Adjusted Average Hazard with Survival Weight (AHSW), RMST, and Event Rate.
}
\details{
It also calculates confidence intervals and p-values using bootstrapped standard errors.
}
\examples{
library(survival)
library(ggplot2)
library(riskRegression)
library(pammtools)
library(adjustedCurves)

# simulate standard survival data with 300 rows
data_1 <- sim_confounded_surv(n=300, max_t=1.1, group_beta=-0.6)
# code the grouping variable as a factor
data_1$group <- as.factor(data_1$group)

# take a look at the first few rows
head(data_1)

#--- Direct standardization via Cox regression model ----
outcome_model <- survival::coxph(Surv(time, event) ~ x1 + x2 + x3 + x4 + x5 + x6 + group,
                                 data=data_1, x=TRUE)
treatment_model <- glm(group ~ x1 + x2 + x3 + x4 + x5 + x6,
                       data=data_1, family=binomial(link="logit"))

adjsurv <- adjustedsurv(data=data_1,
                        variable="group",
                        ev_time="time",
                        event="event",
                        method="direct",
                        outcome_model=outcome_model,
                        conf_int=TRUE,
                        bootstrap=TRUE,
                        n_boot=100)

#--- AHSW ----
adjusted_ahsw(adjsurv, to=0.4, conf_int=TRUE)

}
\references{
Xiong H, Connors J, Schrag D, Uno H. Comparative Effectiveness Research with Average Hazard for Censored Time-to-Event Outcomes: A Numerical Study <arXiv:2407.00709>

Uno H and Horiguchi M. Ratio and difference of average hazard with survival weight: new measures to quantify survival benefit of new therapy. Statistics in Medicine. 2023;1-17. \doi{10.1002/sim.9651}
}
\author{
Hong Xiong, Hajime Uno
}
