% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_ahsw.R
\name{adjusted_ahsw}
\alias{adjusted_ahsw}
\title{Adjusted Average Hazard with Survival Weight (AHSW)}
\usage{
adjusted_ahsw(adjsurv, to, from=0, conf_int=FALSE, conf_level=0.95)
}
\arguments{
\item{adjsurv}{A list object resulting from an adjusted survival analysis
containing survival probabilities (\code{adjsurv$adj}) and optionally bootstrap
data (\code{adjsurv$boot_data}).}

\item{to}{Numeric. The upper time limit (tau) for calculating RMST and AHSW.}

\item{from}{Numeric. Lower time limit for integration (default 0).}

\item{conf_int}{Logical. If TRUE, calculates bootstrap-based confidence intervals.}

\item{conf_level}{Numeric. Confidence level for intervals (default 0.95).}
}
\value{
A list containing:
\item{est}{A data frame with point estimates of RMST, event rate, and AHSW for two groups, their difference, and log difference.}
\item{rmst}{A data frame with bootstrap-based estimates, standard errors, confidence intervals, and p-values for RMST.}
\item{evrt}{Similar results as \code{rmst} but for event rate.}
\item{ahsw}{Similar results as \code{rmst} but for AHSW.}
}
\description{
Calculate Adjusted Average Hazard with Survival Weight (AHSW), RMST, and Event Rate.
}
\details{
It also calculates confidence intervals and p-values using bootstrapped standard errors.
}
\examples{
adjsurv <- list(
  adj = data.frame(
    time = c(0, 0.5, 1.0, 0, 0.5, 1.0),
    group = factor(c(0, 0, 0, 1, 1, 1)),
    surv = c(1.0, 0.80, 0.70, 1.0, 0.90, 0.85)
  )
)

adjusted_ahsw(adjsurv, to = 1, conf_int = FALSE)

}
\references{
Uno H and Horiguchi M. Ratio and difference of average hazard with survival weight: new measures to quantify survival benefit of new therapy. Statistics in Medicine. 2023;1-17. \doi{10.1002/sim.9651}
}
\author{
Hajime Uno, Miki Horiguchi, Hong Xiong
}
